integration test is finished

+++ RestaurantChainApp/RestaurantChainApp/Dtoes/OrderItemDto.cs
@@ -2,8 +2,8 @@
 {
     public class OrderItemDto : Dto
     {
-        public long OrderId { get; set; }
-        public long MenuItemId { get; set; }
+        public int OrderId { get; set; }
+        public int MenuItemId { get; set; }
         public int Amount { get; set; }
         public double Value { get; set; }
     }
